+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsefunction.py
@@ -509,9 +509,9 @@ class EvaluatorResponseFunction(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-08-20T13:19:51.038Z"
+    created: Optional[str] = "2026-08-20T15:20:52.764Z"
 
-    updated: Optional[str] = "2026-08-20T13:19:51.038Z"
+    updated: Optional[str] = "2026-08-20T15:20:52.764Z"
 
     updated_by_id: OptionalNullable[str] = UNSET